pyproject.toml
[project]
name = "hour-2-running-llms-locally-ollama"
version = "0.1.0"
description = "Agentic RAG chatbot for understanding research papers using Ollama LLMs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core dependencies
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-chroma>=0.1.0",
    "ollama>=0.6.0",
    "pandas>=2.3.2",
    "scikit-learn>=1.7.2",
    "transformers>=4.56.2",
    # Document processing
    "pypdf2>=3.0.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    # Vector stores and embeddings
    "chromadb>=0.4.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    # Text processing and chunking
    "nltk>=3.9.0",
    "spacy>=3.7.0",
    "tiktoken>=0.8.0",
    # Async and HTTP support
    "aiohttp>=3.10.0",
    "httpx>=0.27.0",
    # Configuration and utilities
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    # Logging and monitoring
    "loguru>=0.7.0",
    # Optional: Web interface
    "streamlit>=1.39.0",
    "gradio>=4.44.0",
    "perplexityai>=0.12.0",
]
